Move YamlPipelineJobConfiguration
